requirements.txt
boto3~=1.18.55
channels~=3.0.4
channels-redis~=3.3.1
dateparser~=1.1.0
Django~=3.2.8
django-cors-headers~=3.10.0
django-filter~=21.1
djangorestframework~=3.12.4
Flask~=2.0.2
numpy~=1.19.2
pandas~=1.3.3
pika~=1.2.0
psycopg2-binary~=2.9.1
python-Levenshtein~=0.12.2
redis~=3.5.3
schedule~=1.1.0
scipy~=1.7.1
spacy~=3.1.3
statsmodels~=0.13.0
SQLAlchemy~=1.4.25
tensorflow~=2.6.0
thrift~=0.15.0
xlrd~=1.2.0
docker
SALib